edge_server/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pyserial==3.5
pymodbus==3.5.4
python-dotenv==1.0.0
pydantic==2.5.1
pydantic-settings==2.1.0
loguru==0.7.2
pytest==7.4.3
httpx==0.25.2